% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animated_plot.R
\name{animate_fitness_history}
\alias{animate_fitness_history}
\title{Animate Fitness Change Over Generations}
\usage{
animate_fitness_history(fitness_history)
}
\arguments{
\item{fitness_history}{A list containing fitness values for each generation.}
}
\value{
Animation of fitness history
}
\description{
Animate the change in fitness values over generations.
}
\examples{
# example of usage
fitness_history <- list(c(10, 8, 6, 4, 2), c(9, 7, 5, 3, 1))
animate_fitness_history(fitness_history)

}
