% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioGA-package.R
\docType{package}
\name{BioGA-package}
\alias{BioGA}
\alias{BioGA-package}
\title{BioGA: Bioinformatics Genetic Algorithm (BioGA)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Genetic algorithm are a class of optimization algorithms inspired by the process of natural selection and genetics. This package allows users to analyze and optimize high throughput genomic data using genetic algorithms. The functions provided are implemented in C++ for improved speed and efficiency, with an easy-to-use interface for use within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danymukesha.github.io/BioGA/}
  \item \url{https://bioconductor.org/packages/BioGA/}
  \item Report bugs at \url{https://github.com/danymukesha/BioGA/issues}
}

}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
