% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_fitness_cpp}
\alias{evaluate_fitness_cpp}
\title{Function to evaluate fitness using genomic data}
\usage{
evaluate_fitness_cpp(genomic_data, population)
}
\arguments{
\item{genomic_data}{Numeric matrix of genomic data where rows represent genes/features and columns represent samples.}

\item{population}{Numeric matrix representing the population of individuals.}
}
\value{
Numeric vector of fitness scores for each individual.
}
\description{
Function to evaluate fitness using genomic data
}
