% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mutation_cpp}
\alias{mutation_cpp}
\title{Function to mutate the offspring}
\usage{
mutation_cpp(offspring, mutation_rate)
}
\arguments{
\item{offspring}{Numeric matrix representing the offspring.}

\item{mutation_rate}{Probability of mutation for each individual.}
}
\value{
Numeric matrix representing the mutated offspring.
}
\description{
Function to mutate the offspring
}
