% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plottings.R
\name{plot_fitness}
\alias{plot_fitness}
\title{Plot Fitness Values}
\usage{
plot_fitness(fitness_values)
}
\arguments{
\item{fitness_values}{A numeric vector containing fitness values.}
}
\value{
Plot of fitness
}
\description{
Plot the fitness values of the population over generations.
}
\examples{
# example of usage
fitness_values <- c(10, 8, 6, 4, 2)
plot_fitness(fitness_values)

}
