% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plottings.R
\name{plot_fitness_history}
\alias{plot_fitness_history}
\title{Plot Fitness Change Over Generations}
\usage{
plot_fitness_history(fitness_history)
}
\arguments{
\item{fitness_history}{A list containing fitness values for
each generation.}
}
\value{
Plot of fitness history
}
\description{
Plot the change in fitness values over generations.
}
\examples{
# example of usage
fitness_history <- list(c(10, 8, 6, 4, 2), c(9, 7, 5, 3, 1))
plot_fitness_history(fitness_history)

}
