% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_population.R
\name{plot_population}
\alias{plot_population}
\title{Plot Population Distribution}
\usage{
plot_population(population)
}
\arguments{
\item{population}{A numeric matrix containing the population data.}
}
\value{
Plot of population
}
\description{
Plot the distribution of individuals in the population.
}
\examples{
# example of usage
population <- matrix(runif(100), nrow = 10, ncol = 10)
plot_population(population)

}
