% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioGA-package.R
\docType{package}
\name{BioGA-package}
\alias{BioGA}
\alias{BioGA-package}
\title{BioGA: Bioinformatics Genetic Algorithm (BioGA)}
\description{
Genetic algorithm are a class of optimization algorithms inspired by the process of natural selection and genetics. This package allows users to analyze and optimize high throughput genomic data using genetic algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danymukesha.github.io/BioGA/}
}

}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
