% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_population_cpp}
\alias{initialize_population_cpp}
\title{Function to initialize the population from genomic data}
\usage{
initialize_population_cpp(genomic_data, population_size)
}
\arguments{
\item{genomic_data}{Numeric matrix of genomic data where rows represent genes/features and columns represent samples.}

\item{population_size}{Number of individuals in the population.}
}
\value{
Numeric matrix representing the initialized population.
}
\description{
Function to initialize the population from genomic data
}
