% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replacement_cpp}
\alias{replacement_cpp}
\title{Function to replace non-selected individuals in the population}
\usage{
replacement_cpp(population, offspring, num_to_replace)
}
\arguments{
\item{population}{Numeric matrix representing the population of individuals.}

\item{offspring}{Numeric matrix representing the offspring.}

\item{num_to_replace}{Number of individuals to replace.}
}
\value{
Numeric matrix representing the updated population.
}
\description{
Replace non-selected individuals in the population
}
